Pull the avatar menu code out of MainView
...in preparation for showing this menu in the bottom-left/bottom-right
floating menu if enabled.
